moved the order_products migration v1.12.1.1

+++ back-end API/OMUS_api/database/migrations/2023_08_07_121200_create_order_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained('products')->onDelete('cascade');
             $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
-            $table->double('total_price');
+            $table->float('total_price');
             $table->timestamps();
         });
     }
